honeypot/requirements.txt
hpfeeds
twisted
msgpack
pyzmq